fix windows build
ssw.c needed to declare varaibles at the top of code blocks.
Added portable getopt() to the test programs.  KSEQ_INIT() needed to be passed a portable read/_read function name.

+++ apps/parasail_aligner.cpp
@@ -1150,8 +1150,8 @@ inline static void process(
         const char &sentinal,
         const int &cutoff)
 {
-    const int n_children = q.children.size();
-    int child_index = 0;
+    const size_t n_children = q.children.size();
+    size_t child_index = 0;
 
     if (q.lcp < cutoff) return;
